% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocor.R
\name{autocor_lag1}
\alias{autocor_lag1}
\title{Quick Autocorrelation of Lag 1}
\usage{
autocor_lag1(x, ..., check = TRUE)
}
\arguments{
\item{x}{values}

\item{...}{ignored}

\item{check}{check input?}
}
\description{
1D only, inteded for internal use.
}
