% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexual_mann_kendall.R
\name{contextual_mann_kendall}
\alias{contextual_mann_kendall}
\title{Contextual Mann-Kendall Trend Test}
\usage{
contextual_mann_kendall(x, ..., neighbourhood = 2)
}
\arguments{
\item{x}{Raster stack, each layer denoting one timestep. Equidistant steps of time 1.}

\item{neighbourhood}{2:queen, 1:rook, 0:none (classical M-K test)}
}
\description{
2D raster version
}
\references{
Neeti, N. and Eastman J.R. (2011) A Contextual Mann-Kendall Approach for the Assesment of Trend Significance in Image Time Series, \emph{Transactions in GIS}
}
