% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox-stuart-trend.R
\name{cox_stuart}
\alias{cox_stuart}
\title{Cox-Stuart Trend Test}
\usage{
cox_stuart(x, check = TRUE, split = 0.5, alternative = "two.sided")
}
\arguments{
\item{x}{A vector of time series values in order}

\item{check}{check the data. If you give nx2 matrix then this FALSE saves time.}

\item{split}{<= 0.5. Size of the beginning/end of series to use in comparison.}

\item{alternative}{"two.sided", "less" (downwards) or "greater" (upwards trend)}
}
\description{
Cox-Stuart Trend Test
}
\details{
The test is done two tailed, no up/down direction.
}
