% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox-stuart-trend.R
\name{cox_stuart_stack}
\alias{cox_stuart_stack}
\title{Cox-Stuart Trend Test For a Stack}
\usage{
cox_stuart_stack(x, split = 0.5, ...)
}
\arguments{
\item{x}{rasterStack}

\item{split}{See \link{cox_stuart}}

\item{...}{ignored}
}
\description{
Cox-Stuart Trend Test For a Stack
}
\details{
Stack's layers are taken as time. Each cell's timeseries is passed on to \link{cox_stuart}.
}
