% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal_average.R
\name{focal_average}
\alias{focal_average}
\title{Raster Focal Statistics}
\usage{
focal_average(x, ..., neighbourhood = 2)
}
\arguments{
\item{x}{raster or rasterStack}

\item{neighbourhood}{2: queen}
}
\description{
Average value in neighbourhood of each cell.
}
\details{
NA-neighbour values will be ignored and will not increase the denominator.

Edge pixels have the correct denominators.

The result will not fill in NA cells, so if the focal cell value is NA, it will be NA in the output.

The input 'x' will be converted to a rasterStack with 'stack(x)' before calculations.
}
