% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall.R
\name{mann_kendall}
\alias{mann_kendall}
\title{Mann-Kendall Trend Test}
\usage{
mann_kendall(x, ..., est_beta = TRUE, check = TRUE)
}
\arguments{
\item{x}{Time series, either a vector of values or n x 2 matrix of value,time pairs}

\item{est_beta}{Estimate linear coefficient as well?}

\item{check}{check the data. If you give nx2 matrix then this TRUE saves time.}
}
\description{
1D. Assuming time
}
