% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall.R
\name{mann_kendall_stack}
\alias{mann_kendall_stack}
\title{Mann-Kendall Trend Test for Stack}
\usage{
mann_kendall_stack(x, ..., neighbourhood = 0)
}
\arguments{
\item{x}{rasterStack}

\item{...}{passed on to \link{contextual_mann_kendall}}

\item{neighbourhood}{default: 0. See \link{contextual_mann_kendall}}
}
\description{
1D. Assuming time as layers.
}
\details{
Just a wrapper for \link{contextual_mann_kendall} with neighbourhood=0 i.e. no
smoothing applied.
}
