% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-merge-wrapper.R
\name{mergeRaster}
\alias{mergeRaster}
\title{Merge RasterLayers}
\usage{
mergeRaster(x, ..., fun = NULL)
}
\arguments{
\item{x}{list of rasters}

\item{...}{ignored}

\item{fun}{function to apply on the overlapping regions}
}
\description{
Copied from SpaDES.tools which has large requirements.
}
