% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-merge-wrapper.R
\name{mergeRaster_cell}
\alias{mergeRaster_cell}
\title{Merge Rasters Split by row-col}
\usage{
mergeRaster_cell(x, rclist, r0, buffer)
}
\arguments{
\item{x}{list of rasters, split by row-col}

\item{rclist}{row-col metadata of the split}

\item{r0}{raster with original details}

\item{buffer}{size of buffer used for splitting}
}
\description{
Merge Rasters Split by row-col
}
\details{
No checks.
}
