% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarise-p-values.R
\name{p.adjust_raster}
\alias{p.adjust_raster}
\title{Adjust p-values in a Raster}
\usage{
p.adjust_raster(x, method = c("none", "holm"))
}
\arguments{
\item{x}{A raster of p-values}

\item{method}{Type of discovery}

\item{alpha}{nominal level, like 0.05 or 0.01}
}
\value{
a raster-object
}
\description{
Simple wrapper for the p.adjust-function.
}
\details{
See \link{p.adjust} for the method parameters.
}
