% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pettitt-changepoint-test.R
\name{pettitt}
\alias{pettitt}
\title{Pettitt's Test For a Change-Point}
\usage{
pettitt(x, check = TRUE)
}
\arguments{
\item{x}{A vector of time series values in order}

\item{check}{check the data. If you give nx2 matrix then this FALSE saves time.}
}
\description{
Test for a single changepoint in a timeseries.
}
