% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pettitt-changepoint-test.R
\name{pettitt_stack}
\alias{pettitt_stack}
\title{Pettitt's Change-Point Test for Stack}
\usage{
pettitt_stack(x, ...)
}
\arguments{
\item{x}{rasterStack}

\item{...}{passed on to \link{pettitt}}
}
\description{
Assuming time as layers
}
\details{
Stack's layers are taken as time. Each cell's timeseries is passed on to \link{pettitt}.
}
