% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_covariance.R
\name{stack_covariance}
\alias{stack_covariance}
\title{Covariance Matrix of Timeseries RasterStack}
\usage{
stack_covariance(x, neighbourhood = 2, sparse = TRUE, ver = 2)
}
\arguments{
\item{x}{Raster stack.}

\item{neighbourhood}{0: none 1:rook 2:queen (default is 2)}

\item{sparse}{Logical: Use spareMatrix? (default TRUE, highly recommended)}

\item{ver}{leave at 2. (internal development only.)}
}
\description{
Compute the covariance of two timeseries in neighbouring cells.
}
\details{
Compute the covariances between timeserieses in rasterStack cells, limited to neighbouring cells.
}
