% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theil_sen.R
\name{theil_sen}
\alias{theil_sen}
\title{Theil-Sen Slope Estimator}
\usage{
theil_sen(x, t = 1:length(x), ...)
}
\arguments{
\item{x}{Vector of values}

\item{t}{Optional vector of times, numerical}

\item{...}{ignored}
}
\description{
Estimate the slope of a time-series by the
median of all pairwise 1-step slopes.
}
