% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_swail_prewhiten.R
\name{wang_swail_prewhiten_1d}
\alias{wang_swail_prewhiten_1d}
\title{Prewhitening 1D}
\usage{
wang_swail_prewhiten_1d(x, ..., eps = 1e-04, rho_th = 0.05,
  itmax = 20, useC = TRUE)
}
\arguments{
\item{x}{either vector or matrix with c(x, time)}
}
\value{
list with element W containing the whitened series, one shorter than the original.
}
\description{
Simple 1d timeseries: attempt to remove autocorrelation. Assumes AR(1) errors, and 1 step time-grid.
}
\details{
time is ignored, and assumed to be 1-step integers 1:nrow(x).
}
\references{
Wang, X.L. and Swail, V.R. (2000) Changes of Extreme Wave Heights
in Northern Hemisphere Oceans and Related Atmoshperic Circulation Regimes, \emph{Journal of Climate}
}
