% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_swail_prewhiten.R
\name{wang_swail_prewhiten_stack}
\alias{wang_swail_prewhiten_stack}
\title{Prewhitening Raster Stack}
\usage{
wang_swail_prewhiten_stack(x, ...)
}
\arguments{
\item{x}{rasterStack}

\item{...}{Passed on to \link{wang_swail_prewhiten_1d}}
}
\description{
rasterStack version
}
\details{
Apply Wang&Swail prewhitening pixelwise. The method assumes AR(1) auto-correlation.
The resulting stack will have first layer just NA.
}
\references{
Wang, X.L. and Swail, V.R. (2000) Changes of Extreme Wave Heights
in Northern Hemisphere Oceans and Related Atmoshperic Circulation Regimes, \emph{Journal of Climate}
}
\seealso{
\link{wang_swail_prewhiten_1d}
}
