% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall.R
\name{mann_kendall_stack}
\alias{mann_kendall_stack}
\title{Mann-Kendall Trend Test for Stack}
\usage{
mann_kendall_stack(x, ..., est_beta = TRUE)
}
\arguments{
\item{x}{rasterStack}

\item{est_beta}{Estimate linear coefficient as well? Takes extra effort}
}
\description{
1D. Assuming time as layers
}
