% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-merge-wrapper.R
\name{split_calc_wrapper}
\alias{split_calc_wrapper}
\title{Split-Calc-Merge Wrapper}
\usage{
split_calc_wrapper(x, nx, ny, buffer = c(1, 1), fun, ..., dbg = FALSE,
  byrowcol = TRUE)
}
\arguments{
\item{x}{Raster stack}

\item{nx}{Number of x-dimension divisions}

\item{ny}{Number of y-dimension divisions}

\item{buffer}{Extent each sub-raster* by this many cells}

\item{fun}{The function to call per piece. Should return either a raster or stack of identical dimensions as input for it.}

\item{...}{passed to fun in argument list after the piece}

\item{dbg}{Print messages?}

\item{byrowcol}{Split by row-col indices instead of spatial extents? Recommended to keep this TRUE}
}
\description{
Takes a raster stack, splits it into pieces, apply fun for each sub-stack, and then merges together.
}
\details{
The result should be identical to just calling 'fun(x, ...)'.

This function automatically takes advantage of raster-package's multicore setup. See \link{beginCluster} for details how to use it.
}
\note{
If byrowcol=FALSE: a stitching message about different buffers might appear (from 'mergeRaster()'), the (nx,ny) pair
is problematic for 'mergeRaster' and the resulting raster will have artifacts in the splitted edges. Different (nx,ny) is the only way to get around this at the moment.
}
\references{
Neeti, N. and Eastman J.R. (2011) A Contextual Mann-Kendall Approach for the Assesment of Trend Significance in Image Time Series, \emph{Transactions in GIS}
}
