% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-merge-wrapper.R
\name{split_calc_wrapper}
\alias{split_calc_wrapper}
\title{Split-Calc-Merge Wrapper}
\usage{
split_calc_wrapper(x, nx, ny, buffer = c(1, 1), fun, ..., dbg = FALSE)
}
\arguments{
\item{x}{Raster stack}

\item{nx}{passed on to splitRaster}

\item{ny}{passed on to splitRaster}

\item{buffer}{passed on to splitRaster}

\item{fun}{The function to call per piece. Should return either a raster or stack}

\item{...}{passed to fun in argument list after the piece}

\item{dbg}{Print messages?}

\item{neighbourhood}{2:queen, 1:rook, 0:none (classical M-K test)}
}
\description{
Takes a raster stack, splits it to pieces, apply fun for each sub-stack, and then merges together.
}
\details{
The result should be identical to just calling 'fun(x, ...)'.
}
\note{
If at stitching a message about different buffers appears (from 'mergeRaster()'), the (nx,ny) pair
is problematic for 'mergeRaster' and the resulting raster will have artifacts in the splitted edges. Different (nx,ny) is the only way to get around this at the moment. (Should be fixed by devs!)
}
\references{
Neeti, N. and Eastman J.R. (2011) A Contextual Mann-Kendall Approach for the Assesment of Trend Significance in Image Time Series, \emph{Transactions in GIS}
}
